requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.4
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
packaging==21.3
psycopg2-binary==2.8.6
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
sqlparse==0.4.2
wrapt==1.14.0
